+++ sample-project/www/js/index.js
@@ -37,11 +37,9 @@ onDeviceReady: function() {
     document.getElementById("changeUserBtn").addEventListener("click", changeUser);
     document.getElementById("logCustomEventBtn").addEventListener("click", logCustomEvent);
     document.getElementById("logPurchaseBtn").addEventListener("click", logPurchase);
-    document.getElementById("submitFeedbackBtn").addEventListener("click", submitFeedback);
     document.getElementById("setCustomUserAttributeBtn").addEventListener("click", setCustomUserAttribute);
     document.getElementById("setUserPropertiesBtn").addEventListener("click", setUserProperties);
     document.getElementById("launchNewsFeedBtn").addEventListener("click", launchNewsFeed);
-    document.getElementById("launchFeedbackBtn").addEventListener("click", launchFeedback);
     document.getElementById("unsetCustomUserAttributeBtn").addEventListener("click", unsetCustomUserAttribute);
     document.getElementById("setCustomUserAttributeArrayBtn").addEventListener("click", setCustomUserAttributeArray);
     document.getElementById("incrementCustomUserAttributeBtn").addEventListener("click", incrementCustomUserAttribute);
@@ -112,11 +110,6 @@ function logPurchase() {
     AppboyPlugin.logPurchase("testPurchaseWithNullQuantity", 10, "USD");
     AppboyPlugin.logPurchase("testPurchaseWithoutProperties", 1500, "JPY", 2);
     showTextBubble("Logged purchase");
-}
-
-function submitFeedback() {
-    AppboyPlugin.submitFeedback("cordova@test.com", "nice app!", true);
-    showTextBubble("Submitted feedback");
 }
 
 // Appboy User methods
@@ -199,10 +192,6 @@ function requestDataFlush() {
 // Launch functions
 function launchNewsFeed() {
     AppboyPlugin.launchNewsFeed();
-}
-
-function launchFeedback() {
-    AppboyPlugin.launchFeedback();
 }
 
 // News feed functions