+++ src/ios/AppboyPlugin.h
@@ -4,7 +4,6 @@
 
 /*-------Appboy.h-------*/
 - (void) changeUser:(CDVInvokedUrlCommand *)command;
-- (void) submitFeedback:(CDVInvokedUrlCommand *)command;
 - (void) logCustomEvent:(CDVInvokedUrlCommand *)command;
 - (void) logPurchase:(CDVInvokedUrlCommand *)command;
 - (void) disableSdk:(CDVInvokedUrlCommand *)command;
@@ -39,7 +38,6 @@
 
 /*-------Appboy UI-------*/
 - (void) launchNewsFeed:(CDVInvokedUrlCommand *)command;
-- (void) launchFeedback:(CDVInvokedUrlCommand *)command;
 
 /*-------News Feed-------*/
 - (void) getCardCountForCategories:(CDVInvokedUrlCommand *)command;